[kernel] Initial work

+++ src/emu/core/include/core/arm/jit_unicorn.h
@@ -16,15 +16,15 @@ namespace eka2l1 {
 
             void step() override;
 
-            uint64_t get_reg(size_t idx) override;
+            uint32_t get_reg(size_t idx) override;
             uint64_t get_sp() override;
             uint64_t get_pc() override;
-            uint128_t get_vfp(size_t idx) override;
+            uint64_t get_vfp(size_t idx) override;
 
-            void set_reg(size_t idx, uint64_t val) override;
+            void set_reg(size_t idx, uint32_t val) override;
             void set_pc(uint64_t val) override;
             void set_lr(uint64_t val) override;
-            void set_vfp(size_t idx, uint128_t val) override;
+            void set_vfp(size_t idx, uint64_t val) override;
 
             uint32_t get_cpsr() override;
